SourceCode/DoorbellPiService/requirements.txt
Flask>=2.1.2
python-dotenv>=0.20.0
python-decouple>=3.6
requests>=2.28.1
RPi.GPIO>=0.7.0
opencv-python>=4.6.0.66
keyboard>=0.13.5
twilio>=7.12.0
arsenic>=21.8
aioflask>=0.4.0
numpy>=1.23.1